+++ src/main/groovy/io/saagie/plugin/dataops/tasks/projects/ProjectCreatePipelineTask.groovy
@@ -1,4 +1,4 @@
-package io.saagie.plugin.dataops.tasks
+package io.saagie.plugin.dataops.tasks.projects
 
 import io.saagie.plugin.dataops.DataOpsExtension
 import io.saagie.plugin.dataops.clients.SaagieClient
@@ -7,13 +7,13 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.TaskAction
 
-class ProjectUpdatePipelineTask extends DefaultTask {
+class ProjectCreatePipelineTask extends DefaultTask {
     @Input DataOpsExtension configuration
     @Internal SaagieClient saagieClient
 
     @TaskAction
     def createProjectPipeline() {
         saagieClient = new SaagieClient(configuration)
-        logger.quiet(saagieClient.updateProjectPipeline())
+        logger.quiet(saagieClient.createProjectPipelineJob())
     }
 }
