+++ src/main/groovy/io/saagie/plugin/dataops/tasks/projects/ProjectDeletePipelineTask.groovy
@@ -1,4 +1,4 @@
-package io.saagie.plugin.dataops.tasks
+package io.saagie.plugin.dataops.tasks.projects
 
 import io.saagie.plugin.dataops.DataOpsExtension
 import io.saagie.plugin.dataops.clients.SaagieClient
@@ -7,13 +7,13 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.TaskAction
 
-class ProjectStopJobInstanceTask extends DefaultTask {
+class ProjectDeletePipelineTask extends DefaultTask {
     @Input DataOpsExtension configuration
     @Internal SaagieClient saagieClient
 
     @TaskAction
-    def stopPipelineInstance() {
+    def deleteProjectPipeline() {
         saagieClient = new SaagieClient(configuration)
-        logger.quiet(saagieClient.stopJobInstance())
+        logger.quiet(saagieClient.deleteProjectPipeline())
     }
 }
