+++ src/main/groovy/io/saagie/plugin/dataops/tasks/projects/ProjectGetJobInstanceStatus.groovy
@@ -1,4 +1,4 @@
-package io.saagie.plugin.dataops.tasks
+package io.saagie.plugin.dataops.tasks.projects
 
 import io.saagie.plugin.dataops.DataOpsExtension
 import io.saagie.plugin.dataops.clients.SaagieClient
@@ -7,13 +7,13 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.TaskAction
 
-class ProjectArchiveJobTask extends DefaultTask {
+class ProjectGetJobInstanceStatus extends DefaultTask {
     @Input DataOpsExtension configuration
     @Internal SaagieClient saagieClient
 
     @TaskAction
-    def archiveProjectJob() {
+    def getJobInstanceStatus() {
         saagieClient = new SaagieClient(configuration)
-        logger.quiet(saagieClient.archiveProjectJob())
+        logger.quiet(saagieClient.getJobInstanceStatus())
     }
 }
