+++ src/main/groovy/io/saagie/plugin/dataops/tasks/projects/ProjectRunPipelineTask.groovy
@@ -1,4 +1,4 @@
-package io.saagie.plugin.dataops.tasks
+package io.saagie.plugin.dataops.tasks.projects
 
 import io.saagie.plugin.dataops.DataOpsExtension
 import io.saagie.plugin.dataops.clients.SaagieClient
@@ -7,13 +7,13 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.TaskAction
 
-class ProjectStopPipelineInstanceTask extends DefaultTask {
+class ProjectRunPipelineTask extends DefaultTask {
     @Input DataOpsExtension configuration
     @Internal SaagieClient saagieClient
 
     @TaskAction
-    def stopPipelineInstance() {
+    def createProjectJob() {
         saagieClient = new SaagieClient(configuration)
-        logger.quiet(saagieClient.stopPipelineInstance())
+        logger.quiet(saagieClient.runProjectPipeline())
     }
 }
